requirements.txt
requests==2.9.1
python-dotenv==0.4.0
djangorestframework==3.3.3
